requirements.txt
Django==1.9.7
Pillow==3.2.0
Unidecode==0.04.19
Willow==0.3.1
beautifulsoup4==4.4.1
dj-database-url==0.4.1
dj-static==0.0.6
django-modelcluster==2.0
django-taggit==0.18.3
django-toolbelt==0.0.1
django-treebeard==4.0.1
djangorestframework==3.3.3
gunicorn==19.6.0
html5lib==0.9999999
psycopg2==2.6.1
pytz==2016.4
six==1.10.0
static3==0.7.0
wagtail==1.5.2
whitenoise==3.2
